fix(web): open Showcase image attachments in the in-app lightbox
The static Showcase landing demo rendered an image attachment as
<a class="msg-att-img" target="_blank">, so clicking the preview navigated
the browser away to the raw image instead of opening the floating lightbox the
real Chat view uses. Extract Chat's Lightbox into a shared web/src/Lightbox.tsx
and reuse it from both Chat and Showcase; the Showcase image is now a button
that opens the lightbox in place. Download card unchanged.

Adds a failing-first regression test; repoints the existing lightbox layout test
at the shared module; documents the new module in ARCHITECTURE.md (doc-sync).
