+++ ARCHITECTURE.md
@@ -104,7 +104,7 @@ Lets any machine join a server without cloning the repo: `npx @fancyboi999/open-
 - `views/misc.tsx` — Tasks / Inbox / Computers / Search / Settings.
 - `views/Landing.tsx` / `views/Features.tsx` / `views/ProductMock.tsx` + `landing/` — Public marketing pages: landing (`/`) plus features (`/features`) share the warm-editorial skin (tokens all scoped to `.lp-*`, strictly isolated from app editorial skin — no cross-contamination). Landing covers hero / three-pillar / capability cards × 9 / engine / architecture / CTA / footer; features is a static public case showcase with five channel→task→thread demos and a clickable thread panel. Both pages reuse `ProductMock` so the promo surface mirrors the real app shell (rail, channel sidebar, chat, task pill, thread, agent profile, existing avatar system), with no API or live-agent calls. "Enter workspace" navigates to `/s/:slug` when logged in, otherwise to `/login`. Copy and selling points are grounded in verified capabilities.
 - `TaskBoard.tsx` — 5-column kanban (todo / in_progress / in_review / done / closed) + List toggle + **Board layout toggle** (horizontal columns ↔ vertical stack, default horizontal, persisted in `localStorage["open-tag.tasks.boardLayout"]`; horizontal uses min-width columns + horizontal scroll so the narrow embedded channel tasks tab degrades gracefully) + filters + **move interaction** (drag-and-drop *or* click the status pill's menu — interactive controls inside a card `stopPropagation` the `pointerdown` so dnd-kit's sensor doesn't swallow the click; the board fills the pane (`.board-scroll` flex column) and columns are **full-height equal lanes** (`align-items:stretch`, body `flex:1` + internal `overflow-y`), so the *whole* column is the drop target, not just the cards; a `.drop-slot` indicator is appended after the cards instead of overlapping them; moves are optimistic and the card FLIP-animates from its old column to its new one, honoring `prefers-reduced-motion`) + real-time subscription.
-- `Layout/Avatar/Select/PaneEmpty/icons.tsx`, `styles.css` — Shell / avatar / dropdown / centered editorial empty-state (`<PaneEmpty>`, for primary full content panes — secondary sub-section empties keep the inline `.empty`) / icons + UI tokens (EB Garamond + Inter, drawn from DESIGN.md token subset).
+- `Layout/Avatar/Select/PaneEmpty/Lightbox/icons.tsx`, `styles.css` — Shell / avatar / dropdown / centered editorial empty-state (`<PaneEmpty>`, for primary full content panes — secondary sub-section empties keep the inline `.empty`) / shared image lightbox (`Lightbox.tsx`: portal to `document.body`, scroll-zoom + drag-pan + Esc/backdrop-to-close + focus-trap; reused by `views/Chat.tsx` message attachments and the static `views/Showcase.tsx` demo so an image preview opens a floating dialog in place instead of navigating to the raw asset) / icons + UI tokens (EB Garamond + Inter, drawn from DESIGN.md token subset).
 - `trajBuffer.ts` — Pure bounded ring buffer (`appendCapped`, `TRAJ_CAP=300`) for the global Agent Live Trace feed; the memory bound for the in-session trace held in `store.tsx`.
 - `dist/` — App build output, statically served by the server `index.ts`.
