+++ web/src/views/Chat.tsx
@@ -1,5 +1,4 @@
 import { useEffect, useLayoutEffect, useRef, useState, type CSSProperties } from "react";
-import { createPortal } from "react-dom";
 import { useParams, useNavigate, useSearchParams } from "react-router-dom";
 import { useTranslation } from "react-i18next";
 import i18n from "../i18n";
@@ -11,6 +10,7 @@ import { Smile, X, ExternalLink, CheckCircle2, MessageCircle, MoreHorizontal, Li
 const TASK_ICON: Record<string, typeof Circle> = { todo: Circle, in_progress: Play, in_review: Eye, done: CheckCircle2, closed: Ban };
 import { IconWrench, IconFile, IconExternalLink, IconDownload } from "../icons.tsx";
 import { Avatar, resolveAvatar } from "../Avatar.tsx";
+import { Lightbox } from "../Lightbox.tsx";
 import { TaskBoard, ynOptions, ST_LABEL } from "../TaskBoard.tsx";
 import { PaneEmpty } from "../PaneEmpty.tsx";
 import { ChatSkeleton } from "./Skeleton.tsx";
@@ -23,46 +23,6 @@ import { useConfirm, useEscClose } from "../ConfirmModal.tsx";
 const fmtSize = (n?: number) => (!n ? "" : n < 1024 ? n + " B" : n < 1048576 ? (n / 1024).toFixed(1) + " KB" : (n / 1048576).toFixed(1) + " MB");
 const isImage = (m?: string) => !!m && m.startsWith("image/");
 const isVideo = (m?: string) => !!m && m.startsWith("video/");
-
-// Image lightbox: focused media panel with scroll-to-zoom, drag-to-pan, double-click to reset, Esc/backdrop to close
-function Lightbox({ src, alt, onClose }: { src: string; alt: string; onClose: () => void }) {
-  const [scale, setScale] = useState(1);
-  const [pos, setPos] = useState({ x: 0, y: 0 });
-  const drag = useRef<{ x: number; y: number } | null>(null);
-  const closeRef = useRef<HTMLButtonElement>(null);
-  const prevFocus = useRef<HTMLElement | null>(null);
-  useEffect(() => {
-    prevFocus.current = document.activeElement instanceof HTMLElement ? document.activeElement : null;
-    closeRef.current?.focus();
-    const h = (e: KeyboardEvent) => {
-      if (e.key === "Escape") onClose();
-      if (e.key === "Tab") {
-        e.preventDefault();
-        closeRef.current?.focus();
-      }
-    };
-    window.addEventListener("keydown", h);
-    return () => {
-      window.removeEventListener("keydown", h);
-      prevFocus.current?.focus();
-    };
-  }, [onClose]);
-  return createPortal(
-    <div className="lightbox-bg" role="dialog" aria-modal="true" aria-label={alt} onClick={onClose} onWheel={(e) => { setScale((s) => Math.min(8, Math.max(1, s - e.deltaY * 0.0016 * s))); }}>
-      <button ref={closeRef} className="lightbox-x" onClick={onClose} aria-label={i18n.t("chat.close")}><X size={20} /></button>
-      <div className="lightbox-panel" onClick={(e) => e.stopPropagation()}>
-        <img src={src} alt={alt} className="lightbox-img" draggable={false}
-          style={{ transform: `translate(${pos.x}px,${pos.y}px) scale(${scale})`, cursor: scale > 1 ? (drag.current ? "grabbing" : "grab") : "zoom-in" }}
-          onClick={(e) => { e.stopPropagation(); if (scale === 1) setScale(2); }}
-          onDoubleClick={(e) => { e.stopPropagation(); setScale(1); setPos({ x: 0, y: 0 }); }}
-          onMouseDown={(e) => { if (scale > 1) { e.preventDefault(); drag.current = { x: e.clientX - pos.x, y: e.clientY - pos.y }; } }}
-          onMouseMove={(e) => { if (drag.current) setPos({ x: e.clientX - drag.current.x, y: e.clientY - drag.current.y }); }}
-          onMouseUp={() => { drag.current = null; }} onMouseLeave={() => { drag.current = null; }} />
-      </div>
-    </div>,
-    document.body,
-  );
-}
 
 // Message attachments: images shown inline with lightbox, videos shown with inline player; other files shown as a file card. Videos that fail to play fall back to a download card.
 function AttCard({ a, url }: { a: Att; url: string }) {
